Database Token Driver: Enhanced token security
The token field in the tokens table has been changed from a string to a char of length 128. This is to accommodate hashed tokens, which are now being used for increased security. The DatabaseTokenDriver class has been updated accordingly. A new method, hashToken(), has been added to this class, which hashes tokens using the SHA-512 algorithm before they are stored or queried in the database. Additionally, random_bytes() function is now used instead of Str::random() for generating tokens, providing better randomness and security.